Sotib olish bo'limini boshladim

+++ MOCK_DATA.sql
@@ -7,7 +7,7 @@ CREATE TABLE Medicine_items(
     name VARCHAR(64) UNIQUE,
     produced_time DATE,
     end_time DATE,
-    expiration_date float,
+    expiration_date INT,
     price INT,
     count INT
 );
